+++ Plugins/BetterRoleColors/BetterRoleColors.plugin.js
@@ -4,7 +4,7 @@ class BetterRoleColors {
 	getName() { return "BetterRoleColors"; }
 	getShortName() { return "BRC"; }
 	getDescription() { return "Adds server-based role colors to typing, voice, popouts, modals and more! Support Server: bit.ly/ZeresServer"; }
-	getVersion() { return "0.7.2"; }
+	getVersion() { return "0.7.3"; }
 	getAuthor() { return "Zerebos"; }
 
 	constructor() {
@@ -58,16 +58,8 @@ class BetterRoleColors {
 		this.RelationshipStore = DiscordModules.RelationshipStore;
 		this.PopoutWrapper = InternalUtilities.WebpackModules.findByUniqueProperties(['Positions', 'Animations']);
 		this.VoiceUser = InternalUtilities.WebpackModules.find(m => typeof(m) === "function" && m.List);
-		this.UserPopout = InternalUtilities.WebpackModules.find(m => {
-			try { return m.displayName == "FluxContainer(Component)" && !(new m()); }
-			catch (e) { return e.toString().includes("user"); }
-		});
-		this.UserModal = InternalUtilities.WebpackModules.find(m => {
-			try {
-				return m.modalConfig && m.prototype.render().type.displayName == "FluxContainer(Component)";
-			}
-			catch (err) {return false;}
-		});
+		this.UserPopout = DiscordModules.UserPopout;
+		this.UserModal = DiscordModules.UserProfileModal;
 		this.AuditLogItem = InternalUtilities.WebpackModules.find(m => m.prototype && m.prototype.renderPermissionUpdate);
 		this.TypingUsers = InternalUtilities.WebpackModules.find(m => {
 			try { return m.displayName == "FluxContainer(t)" && !(new m({channel: 0})); }